+++ app/src/main/java/mx/itesm/sxto/example/proyectoequipo4/model/SismosDatos.kt
@@ -12,6 +12,6 @@ data class SismosDatos(
     @SerializedName("magnitud")
     val magnitud: Double,
     @SerializedName("profundidad")
-    val profundidad: Double
+    val profundidad: Int
 
 )
